+++ FE/src/components/Bottomsheet/KidInfoBox.jsx
