+++ FE/src/components/Bottomsheet/kidInfoBox.module.scss
@@ -1,4 +1,4 @@
-@import "../../../styles/fontStyle.scss";
+@import "@/styles/fontStyle.scss";
 
 .content {
     display: flex;
