+++ FE/src/components/Complete/Complete.module.scss
@@ -1,4 +1,4 @@
-@import "../../../styles/fontStyle.scss";
+@import "@/styles/fontStyle.scss";
 
 .wrapper {
     display: flex;
