+++ FE/src/components/Footer/Footer.module.scss
