+++ FE/src/components/Input/NumericSelector.module.scss
