+++ FE/src/components/Loader/Loader.module.scss
