+++ FE/src/pages/DriverDetail/DriverDetail.jsx
@@ -1,7 +1,7 @@
 import { useLoaderData, useLocation, useNavigate } from "react-router-dom";
-import { Footer } from "../../components/common/Footer/Footer";
-import { Header } from "../../components/common/Header/Header";
-import { getDriverDetail, postSubscribe } from "../../service/api";
+import { Footer } from "@/components/Footer/Footer";
+import { Header } from "@/components/Header/Header";
+import { getDriverDetail, postSubscribe } from "@/service/api";
 import styles from "./DriverDetail.module.scss";
 
 export default function DriverDetail() {
