+++ FE/src/pages/Onboarding/Onboarding.jsx
@@ -3,7 +3,7 @@ import logoIcon from "@/assets/logo.svg";
 import truckIcon from "@/assets/truck.png";
 import userIcon from "@/assets/user_icon.svg";
 import { Type } from "./Type/Type";
-import { getAccessToken } from "../../utils/auth";
+import { getAccessToken } from "@/utils/auth";
 import { redirect } from "react-router-dom";
 
 export default function Onboarding() {
