+++ FE/src/service/api.js
@@ -1,6 +1,6 @@
-import { BASE_URL } from "../constants/constants";
-import { logout } from "../pages/Login/Login";
-import { getAccessToken, getRefreshToken } from "../utils/auth";
+import { BASE_URL } from "@/constants/constants";
+import { logout } from "@/pages/Login/Login";
+import { getAccessToken, getRefreshToken } from "@/utils/auth";
 
 /**
  * @param {Object} props
